fix(sdk): share truncated history backfill across frontend streams
