+++ libs/sdk/src/ui/utils.ts
@@ -1,3 +1,6 @@
+import type { Client } from "../client.js";
+import type { ThreadState } from "../schema.js";
+
 /**
  * Returns true when `onFinish` declares at least one parameter and therefore
  * needs the server-fetched thread head. A zero-arity `onFinish` is treated as
@@ -33,4 +36,44 @@ export async function* filterStream<T, TReturn>(
     if (done) return value as TReturn;
     if (filter(value)) yield value as T;
   }
+}
+
+export function fetchHistory<StateType extends Record<string, unknown>>(
+  client: Client,
+  threadId: string,
+  options?: { limit?: boolean | number }
+): Promise<ThreadState<StateType>[]> {
+  if (options?.limit === false) {
+    return client.threads.getState<StateType>(threadId).then((state) => {
+      if (state.checkpoint == null) return [];
+      return [state];
+    });
+  }
+
+  const limit = typeof options?.limit === "number" ? options.limit : 10;
+  return client.threads
+    .getHistory<StateType>(threadId, { limit })
+    .then(async (history) => {
+      if (history.length < limit) return history;
+
+      const result = [...history];
+      let before = history.at(-1)?.checkpoint?.checkpoint_id;
+
+      while (before != null) {
+        const page = await client.threads.getHistory<StateType>(threadId, {
+          limit,
+          before: before as any,
+        });
+        if (page.length === 0) break;
+
+        result.push(...page);
+        if (page.length < limit) break;
+
+        const nextBefore = page.at(-1)?.checkpoint?.checkpoint_id;
+        if (nextBefore == null || nextBefore === before) break;
+        before = nextBefore;
+      }
+
+      return result;
+    });
 }
